+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 - Update PROMPT_DESCRIBE_PACKAGE
 - Update common instructions about prompt reply language
+- Make sure that projectName is compliant with the format of an environment variable
 
 ## [5.37.0] 2025-05-22